% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd-gras.R
\name{nd_gras}
\alias{nd_gras}
\title{N-dimensional Generalized RAS (nD-GRAS) method}
\usage{
nd_gras(source, constraints, ..., tolerance = 1e-10, max_iterations = 1000)
}
\arguments{
\item{source}{An array to be adjusted.}

\item{constraints}{A list of constraints created by \code{\link[=nd_gras_constraint]{nd_gras_constraint()}}.}

\item{...}{Additional arguments (currently unused).}

\item{tolerance}{A numeric value indicating the convergence tolerance. Default is \code{1e-10}.}

\item{max_iterations}{An integer indicating the maximum number of iterations. Default is \code{1000}.}
}
\value{
A list containing the following components:
\item{target}{The adjusted n-dimensional array.}
\item{margins}{A list of margins used in the constraints.}
\item{multipliers}{A list of multipliers for each constraint.}
\item{iterations}{The number of iterations performed.}
\item{converged}{A logical value indicating whether the algorithm converged.}
}
\description{
N-dimensional Generalized RAS (nD-GRAS) method
}
