% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd-gras-constraint.R
\name{nd_gras_constraint}
\alias{nd_gras_constraint}
\title{Create a constraint for the nD-GRAS method}
\usage{
nd_gras_constraint(margin, target)
}
\arguments{
\item{margin}{An integer vector indicating the dimensions to constrain.}

\item{target}{An array (or vector) representing the target marginal sums.}
}
\value{
A list containing the margin and target.
}
\description{
Create a constraint for the nD-GRAS method
}
